requirements/train.txt
accelerate>=1.4.0
bitsandbytes>=0.45.0
datasets>=3.2.0
modelscope>=1.24.0
peft>=0.14.0
pyyaml>=6.0.2
safetensors>=0.5.3
sentencepiece>=0.2.0
tensorboard>=2.19.0
transformers>=4.51.0

